# sponge/spongejs/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "title": "SpongeJS Blockchain 🧱⛓",
    "description": "General purpose blockchain",
    "version": "1.0.0"
  },
  "host": "localhost:5001",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/network/connect": {
      "post": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
...
  }
}